Correcting direct assignment information.
I am correcting the information on the Direct licenses coexist with group licenses. This information indicates that you are able to modify the user directly even if a group license assignment targets the user. If there is a group license assignment it does not allow the features to be modified directly they are all grayed out. We can however assign another group to the user with the changes and that will apply to the user. Clarifying that in the document.